requirements.txt
openai==1.12.0
langchain==0.1.0
langchain-openai==0.0.5
faiss-cpu==1.7.4
pandas==2.3.1
tqdm==4.67.1
python-dotenv==1.0.0